view orders, change password, forgot password

+++ HSF302_Group3_Project/src/main/java/clothes/hsf302_group3_project/dto/response/CategoryDTO.java
@@ -11,11 +11,10 @@ import java.util.List;
 @NoArgsConstructor
 @Getter
 @Setter
-
 public class CategoryDTO {
-    private Long id;
+    private Integer id;
     private String name;
     private List<ProductDTO> products;
-    private Integer productCount;
+    private int productCount;
 
 }
